Integrate inventory into the configuration
