+++ setup.cfg
@@ -29,3 +29,7 @@ where = src
 
 [options.package_data]
 * = src/faros_config/ui/static/*, src/faros_config/ui/templates/*
+
+[options.entry_points]
+console_scripts =
+    faros-inventory = faros_config.inventory.cli:main
